cleanup controllers and gulp file
