+++ package.json
@@ -26,7 +26,6 @@
     "axios": "^0.18.0",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
-    "chai": "^4.1.2",
     "compression": "^1.7.1",
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
@@ -70,10 +69,8 @@
     "babel-preset-stage-2": "^6.24.1",
     "babel-register": "^6.26.0",
     "browser-sync": "^2.23.5",
-    "del": "^3.0.0",
     "gulp": "^3.9.1",
     "gulp-babel": "^7.0.0",
-    "gulp-load-plugins": "^1.5.0",
     "gulp-newer": "^1.1.0",
     "gulp-nodemon": "^2.0.6",
     "gulp-pug": "^3.3.0",
@@ -81,7 +78,6 @@
     "gulp-sourcemaps": "^1.6.0",
     "gulp-util": "^3.0.7",
     "jest": "^21.2.1",
-    "run-sequence": "^2.2.1",
     "sequelize-cli": "^3.2.0",
     "webpack": "^3.10.0",
     "webpack-stream": "^4.0.0"
